MODULARIZE code
