+++ package.json
@@ -4,7 +4,8 @@
   "description": "weather app to play with OpenWeatherMap API",
   "main": "webpack.config.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "webpack",
+    "watch": "webpack --watch"
   },
   "repository": {
     "type": "git",
@@ -31,9 +32,6 @@
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.17.2",
     "webpack": "^5.4.0",
-    "webpack-cli": "^4.2.0",
-    "browser-sync": "^2.18.8",
-    "gulp": "^3.9.1",
-    "gulp-sass": "^3.1.0"
+    "webpack-cli": "^4.2.0"
   }
 }
